# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Street Burger API",
        "description": "Street Burger Mobile Application Backend API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/auth/signup": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Register a new user",
                "operationId": "App\\Http\\Controllers\\AuthController::signup",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "phone_number",
                                    "password",
                                    "name"
                                ],
                                "properties": {
                                    "phone_number": {
                                        "type": "string",
                                        "example": "0771234567"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password123"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "john@example.com"
                                    },
                                    "date_of_birth": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "1990-01-01"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User registered successfully"
                    },
                    "400": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Login with phone number and password",
                "operationId": "App\\Http\\Controllers\\AuthController::login",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "phone_number",
                                    "password"
                                ],
                                "properties": {
                                    "phone_number": {
                                        "type": "string",
                                        "example": "0771234567"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login successful"
                    },
                    "401": {
                        "description": "Invalid credentials"
                    }
                }
            }
        },
        "/api/auth/generatetoken/{phoneNumber}": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Generate JWT token for a phone number",
                "operationId": "App\\Http\\Controllers\\AuthController::generateToken",
                "parameters": [
                    {
                        "name": "phoneNumber",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Token generated"
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            }
        },
        "/api/auth/delete/account": {
            "delete": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Delete user account",
                "operationId": "App\\Http\\Controllers\\AuthController::deleteAccount",
                "responses": {
                    "200": {
                        "description": "Account deleted successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/menu/items": {
            "get": {
                "tags": [
                    "Menu"
                ],
                "summary": "Get all menu items",
                "operationId": "App\\Http\\Controllers\\MenuController::getAllItems",
                "responses": {
                    "200": {
                        "description": "List of items"
                    }
                }
            },
            "post": {
                "tags": [
                    "Menu"
                ],
                "summary": "Add new item (Admin only)",
                "operationId": "App\\Http\\Controllers\\MenuController::addItem",
                "responses": {
                    "201": {
                        "description": "Item created"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/menu/items/{categoryId}": {
            "get": {
                "tags": [
                    "Menu"
                ],
                "summary": "Get items by category",
                "operationId": "App\\Http\\Controllers\\MenuController::getItemsByCategory",
                "parameters": [
                    {
                        "name": "categoryId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of items in category"
                    }
                }
            }
        },
...
    },
    "security": [
        []
    ]
}